style: auth directory user 하위로 이동
