+++ src/main/kotlin/pmeet/pmeetserver/user/service/oauth/NaverOauthService.kt
@@ -1,4 +1,4 @@
-package pmeet.pmeetserver.auth.service.oauth
+package pmeet.pmeetserver.user.service.oauth
 
 import kotlinx.coroutines.reactor.awaitSingle
 import org.springframework.beans.factory.annotation.Value
@@ -10,19 +10,20 @@ import org.springframework.web.reactive.function.client.bodyToMono
 import pmeet.pmeetserver.user.dto.UserInfo
 
 @Service
-class KakaoOauthService(
+class NaverOauthService(
 
   private val webClient: WebClient,
-  @Value("\${kakao.oauth.client-id}") private val clientId: String,
-  @Value("\${kakao.oauth.redirect-uri}") private val redirectUri: String
+  @Value("\${naver.oauth.client-id}") private val clientId: String,
+  @Value("\${naver.oauth.client-secret}") private val clientSecret: String
 ) {
-  suspend fun getAccessToken(code: String): String {
+  suspend fun getAccessToken(code: String, state: String): String {
     val responseBody: Map<String, Any> = webClient.post()
-      .uri("https://kauth.kakao.com/oauth/token")
+      .uri("https://nid.naver.com/oauth2.0/token/")
       .contentType(MediaType.APPLICATION_FORM_URLENCODED)
       .body(BodyInserters.fromFormData("code", code)
         .with("client_id", clientId)
-        .with("redirect_uri", redirectUri)
+        .with("client_secret", clientSecret)
+        .with("state", state)
         .with("grant_type", "authorization_code"))
       .retrieve()
       .bodyToMono<Map<String, Any>>()
@@ -31,26 +32,24 @@ class KakaoOauthService(
   }
 
   suspend fun getProfile(accessToken: String): UserInfo {
-    val kakaoResponse = webClient.post()
-      .uri("https://kapi.kakao.com/v2/user/me")
-      .contentType(MediaType.APPLICATION_FORM_URLENCODED)
+    val NaverResponse = webClient.post()
+      .uri("https://openapi.naver.com/v1/nid/me")
       .header("Authorization", "Bearer " + accessToken)
       .retrieve()
-      .bodyToMono(KakaoResponse::class.java)
+      .bodyToMono(NaverResponse::class.java)
       .awaitSingle()
-    return UserInfo(kakaoResponse.kakao_account.profile.nickname, kakaoResponse.kakao_account.email)
+    return UserInfo(NaverResponse.response.name, NaverResponse.response.email)
   }
 }
 
-data class KakaoResponse(
-  val kakao_account: KakaoAccount
+data class NaverResponse(
+  val resultcode: String,
+  val message: String,
+  val response: NaverAccount
 )
 
-data class KakaoAccount(
+data class NaverAccount(
+  val id: String,
   val email: String,
-  val profile: KakaoProfile
-)
-
-data class KakaoProfile(
-  val nickname: String
+  val name: String
 )
